BREAKING CHANGES!
Previously, the "current seed" is actually the next seed. This caused a
problem that if you really enter the "current seed", you'll be put on
track B because previously the table was one step ahead, giving an
offset.

It's done so because that's the seed and track I was given.

Now we try to match the "current seed", meaning that we advance the
seed once from the current seed, and then fill the table. This has a few
consequence:

0. All the old links will no longer be accurate. Old links will lose
   one roll, and start from track B. This is because the new table will
   advance the seed once, moving you to track B and lose one roll.
   In order to recover from this, you could just roll once, and now
   it'll just match track B. In order to fully move back to track A,
   just roll and click on the rolled link.
0. The seed seeker requires a similar change. Previously, it's returning
   the next seed right after the last roll. Now it'll just return the
   last seed used by the last roll. This feels weird because that seed
   should have already been used. However, this matches the behaviour of
   real data saved in the game, while it could be confusing, it's more
   accurate about the state.