Now the seeker will try to detect duplicated rare:
So the way the game giving you another rare cat is reducing the
slot code by 1, assuming it'll cycle to max slot when the
slot code is 0. For example, double Rover cats will give you
Rover and then Pogo, because Rover is 0 and Pogo is max.

We did two things to cover this. One is that we detect if there's
a pattern looks like this. If there is, transform the pick to a
dual, which contains both the original cat, and the potential
duplicated cat.

The matching shall first try to match against the original cat,
and if it doesn't match, try the second one. This is done by (<|>).

Secondly, if the first cat is ever a rare cat, it could be a
duplicate after all. We don't know this but for the sake of ease of
using, we just assume it could be one. So if the first cat is a
rare cat, it's always a dual.

This is tested with a few regular duplicated rare cats, and also
double Rover cats, along with having the duplicated cat as the
first rare cat.