Properly parse the event data
So it turned out that there's a field to indicate which offset
we should be using. The index for the offset is 9. Then following it,
for each 14 fields that's a pool data.

Following this pattern, we could properly find the platinum
gacha probability. This also fixed a few probability issues
in the JP events. e.g. Epicfest should be 9% uber, not 5%.

This also means that we don't need to make platinum a special case
when rolling a cat now.