Show the starting seed in seek result:
We introduce a new concept called next_seed, which is only used
when showing the starting seed in seek result.

Details:

This is a special seed indicates that we don't want to
advance the seed for the first roll. In Gacha#current_seed_mode!,
we advance the seed once to simulate the actual states of the game.
That is, the game will advance the seed before rolling.

This is all good, but this also means that we do not know the
very first seed from the seek result, because we can't roll back
the very first seed to see what will roll into the current seed.
Here the `next_seed` indicates that we should treat this seed
as the next seed, therefore we don't advance the first seed.
Thus in the controller we do:

    gacha.current_seed_mode! if next_seed.zero?

When there's no next seed, we enter current seed mode to advance
the first seed. Otherwise, we just use the next seed as the first one.
This is a special case for the very first seek result. Therefore
we don't provide this as an option with an input. This will only
be used in seek_result.erb. After the user clicks on the other links,
they go back to the normal current seed mode because now we know
what exactly is the current seed.