Preserve as much as arguments if possible on form
This way each option is self-contained, we don't need to
worry about putting hidden input in one page or the other,
the form should take care itself.

The downside is that now we need to control the display
on the form directly, rather than on the page. Which is fine
for now. Trade off as always.